🔥 chore: rename directory util to utils
