+++ src/app/middlewares/validateRequest.ts
@@ -1,18 +1,18 @@
 import { ObjectSchema } from "joi";
 import createError from "http-errors";
 
-import { Logger as log } from "../../util/Logger";
+import { Logger as log } from "../../utils/Logger";
 
 import type { Request, Response, NextFunction } from "express";
 
-const validateRequest = (schema: ObjectSchema, property: string) => {
+export const validateRequest = (schema: ObjectSchema, property: string) => {
   return (req: Request, res: Response, next: NextFunction) => {
     const { error } = schema.validate(req[property as keyof typeof req], {
       allowUnknown: true,
     });
 
     if (error) {
-      log.serverError("Schema validation faild.");
+      log.serverError("Schema validation failed.");
 
       next(createError(400, error.message));
     }
@@ -20,5 +20,3 @@ const validateRequest = (schema: ObjectSchema, property: string) => {
     next();
   };
 };
-
-export default validateRequest;
