+++ src/app/middlewares/verifyGithubSignature.ts
@@ -2,7 +2,7 @@ import createError from "http-errors";
 import { createHmac } from "crypto";
 
 import Config from "../../config";
-import { Logger as log } from "../../util/Logger";
+import { Logger as log } from "../../utils/Logger";
 
 import type { RequestHandler } from "express";
