+++ src/app/routes/index.ts
@@ -2,16 +2,15 @@ import { Router } from "express";
 
 import loginRouter from "./login";
 import usersRouter from "./users";
-import projectsRouter from "./projects";
-// import reposRouter from "./repos";
+import { projectsRouter } from "./projects";
 
 const routes = (): Router => {
   const router = Router();
 
   loginRouter(router);
   usersRouter(router);
-  projectsRouter(router);
-  // reposRouter(router);
+
+  router.use("/projects", projectsRouter);
 
   return router;
 };
