+++ src/domains/cmsService/createDomain.ts
@@ -1,9 +1,9 @@
 import { Route53 } from "../../infrastructure/aws";
-import { Logger as log } from "../../util/Logger";
+import { Logger as log } from "../../utils/Logger";
 import Config from "../../config";
 import { CMS_MESSAGE } from "../../config/constants";
 
-export async function createDomain({ subdomain }: any) {
+export async function createDomain({ subdomain }: any): Promise<string> {
   const route53 = new Route53();
   const recordChangeInfo = await route53.createARecord({
     recordName: `${subdomain}.${Config.SERVER_URL}`,
